# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Talk Proposals API",
        "description": "API documentation for the Talk Proposals system. This API allows speakers to submit talk proposals, reviewers to review them, and admins to manage the proposal status.",
        "contact": {
            "email": "support@talkproposals.test"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://api.talkproposals.test/api",
            "description": "Local Development Server"
        },
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server (Alternative)"
        },
        {
            "url": "http://api.talkproposals.test"
        }
    ],
    "paths": {
...
        "/reviews/rating-options": {
            "get": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Get rating options",
                "description": "Returns all available rating options (1-5 and 10) with their labels.",
                "operationId": "1c1d3a7c76b2cc5e5b5c4b4c7e009c6f",
                "responses": {
                    "200": {
                        "description": "Rating options retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Rating options retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "ratings": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "value": {
                                                                "type": "integer",
                                                                "example": 5
                                                            },
                                                            "label": {
                                                                "type": "string",
                                                                "example": "5 - Excellent"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/proposals/{proposalId}/reviews": {
            "get": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Get reviews for a proposal",
                "description": "Retrieves all reviews for a specific proposal, ordered by most recent first.",
                "operationId": "9069365518540905ebc373cff24f514e",
                "parameters": [
                    {
                        "name": "proposalId",
                        "in": "path",
                        "description": "Proposal ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reviews retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Reviews retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "reviews": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Review"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Proposal not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Create a review",
                "description": "Creates a new review for a proposal. Each reviewer can only review a proposal once. Rating must be 1, 2, 3, 4, 5, or 10.",
                "operationId": "c82f32bb02b4e10b083644e094b130d8",
                "parameters": [
                    {
                        "name": "proposalId",
                        "in": "path",
                        "description": "Proposal ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "rating"
                                ],
                                "properties": {
                                    "rating": {
                                        "description": "Rating value (1-5 or 10)",
                                        "type": "integer",
                                        "enum": [
                                            1,
                                            2,
                                            3,
                                            4,
                                            5,
                                            10
                                        ],
                                        "example": 5
                                    },
                                    "comment": {
                                        "description": "Optional review comment",
                                        "type": "string",
                                        "example": "Great proposal!",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Review created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Review created successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "review": {
                                                    "$ref": "#/components/schemas/Review"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized or duplicate review"
                    },
                    "404": {
                        "description": "Proposal not found"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john@example.com"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "speaker",
                            "reviewer",
                            "admin"
                        ],
                        "example": "speaker"
                    }
                },
                "type": "object"
            },
...
            "Review": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "rating": {
                        "type": "integer",
                        "enum": [
                            1,
                            2,
                            3,
                            4,
                            5,
                            10
                        ],
                        "example": 5
                    },
                    "comment": {
                        "type": "string",
                        "example": "Great proposal!",
                        "nullable": true
                    },
                    "reviewer": {
                        "$ref": "#/components/schemas/User"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum SPA authentication using session cookies",
                "scheme": "session"
            }
        }
    },
...
}